Moving onDisconnect to shell
